code review changes
remove <type>jar</type>
remove erroneous spaces
remove explicit failed test validation
remove explicit readonly=false checkbox
remove spring-test
fix help
DRY Move CustomPrefixBlobStoreProvider into its own class
