+++ src/main/resources/io/jenkins/plugins/artifact_manager_jclouds/s3/S3BlobStoreConfig/config.jelly
@@ -21,13 +21,13 @@
             <f:textbox/>
         </f:entry>
         <f:entry title="${%Use Path Style URL}" field="usePathStyleUrl">
-            <f:checkbox readonly="false"/>
+            <f:checkbox/>
         </f:entry>
         <f:entry title="${%Use Insecure HTTP}" field="useHttp">
-            <f:checkbox readonly="false"/>
+            <f:checkbox/>
         </f:entry>
         <f:entry title="${%Disable Session Token}" field="disableSessionToken">
-            <f:checkbox readonly="false" />
+            <f:checkbox/>
         </f:entry>
         <f:validateButton title="Validate S3 Bucket configuration" progress="Validate..." method="validateS3BucketConfig"
                           with="container,prefix,useHttp,usePathStyleUrl,disableSessionToken,customEndpoint,customSigningRegion"/>
