+++ src/main/resources/io/jenkins/plugins/artifact_manager_jclouds/s3/S3BlobStoreConfig/help-usePathStyleUrl.html
@@ -2,5 +2,5 @@
 <br />
 When this option is disabled URLs will be formatted https://bucketname.s3.amazonaws.com/key/path<br />
 When this option is enabled URLs will be formatted https://s3.amazonaws.com/bucketname/key/path<br />
-Typically, you would want this enabled when you're using the real S3 but disabled when using an S3 compatible service.
+Typically, you would only ever enable this option when using an S3 compatible service (not AWS S3) that only supported path style URLs.
 </div>
